requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-multipart==0.0.20
sse-starlette==2.2.1

# HTTP clients
aiohttp==3.11.11
aiohttp-socks==0.10.1
httpx==0.28.1

# HTML parsing
beautifulsoup4==4.12.3
lxml==5.3.0

# Phone validation
phonenumbers==8.13.52

# Neo4j
neo4j==5.27.0

# Graph / ML
torch==2.5.1
torch-geometric==2.6.1
torch-scatter==2.1.2
torch-sparse==0.6.18
networkx==3.4.2
numpy==2.2.1
scikit-learn==1.6.1

# Scheduling
apscheduler==3.10.4

# Visualization data
plotly==6.0.0

# Environment
python-dotenv==1.0.1

# Hunter.io client
pyhunter==1.7

# Utilities
tenacity==9.0.0
